完成供应商与地址的关联管理

+++ ocean-provider/ocean-imes/ocean-imes-basic/src/main/java/com/fantechs/provider/imes/basic/service/impl/SmtSupplierServiceImpl.java
@@ -10,6 +10,7 @@ import com.fantechs.common.base.entity.basic.search.SearchSmtSupplier;
 import com.fantechs.common.base.entity.security.SysUser;
 import com.fantechs.common.base.exception.BizErrorException;
 import com.fantechs.common.base.support.BaseService;
+import com.fantechs.common.base.utils.ClassCompareUtil;
 import com.fantechs.common.base.utils.CurrentUserInfoUtils;
 import com.fantechs.common.base.utils.StringUtils;
 import com.fantechs.provider.imes.basic.mapper.SmtAddressMapper;
@@ -68,23 +69,20 @@ public class SmtSupplierServiceImpl  extends BaseService<SmtSupplier> implements
         int i = smtSupplierMapper.insertUseGeneratedKeys(record);
 
         List<SmtAddressDto> address = record.getList();
-        List<SmtSupplierAddress> supplierAddresses = new ArrayList<>();
-
-        //将新增的地址添加到地址表中
-        smtAddressMapper.insertList(address);
-        //将新增的地址与供应商进行绑定
-        for (SmtAddressDto smtAddress : address) {
-            System.out.println(smtAddress.getAddressId());
-            System.out.println(record.getSupplierId());
-
-            SmtSupplierAddress smtSupplierAddress = new SmtSupplierAddress();
-            smtSupplierAddress.setIfDefault(smtAddress.getIfDefault());
-            smtSupplierAddress.setSupplierId(record.getSupplierId());
-            smtSupplierAddress.setAddressId(smtAddress.getAddressId());
-            supplierAddresses.add(smtSupplierAddress);
+        if (address.size()!=0){
+            List<SmtSupplierAddress> supplierAddresses = new ArrayList<>();
+            //将新增的地址添加到地址表中
+            smtAddressMapper.insertList(address);
+            //将新增的地址与供应商进行绑定
+            for (SmtAddressDto smtAddress : address) {
+                SmtSupplierAddress smtSupplierAddress = new SmtSupplierAddress();
+                smtSupplierAddress.setIfDefault(smtAddress.getIfDefault());
+                smtSupplierAddress.setSupplierId(record.getSupplierId());
+                smtSupplierAddress.setAddressId(smtAddress.getAddressId());
+                supplierAddresses.add(smtSupplierAddress);
+            }
+            smtSupplierAddressMapper.insertList(supplierAddresses);
         }
-
-        smtSupplierAddressMapper.insertList(supplierAddresses);
 
         return i;
     }
@@ -93,9 +91,9 @@ public class SmtSupplierServiceImpl  extends BaseService<SmtSupplier> implements
     @Transactional(rollbackFor = Exception.class)
     public int update(SmtSupplier entity) {
         SysUser currentUser = CurrentUserInfoUtils.getCurrentUserInfo();
-//        if(StringUtils.isEmpty(currentUser)){
-//            throw new BizErrorException(ErrorCodeEnum.UAC10011039);
-//        }
+        if(StringUtils.isEmpty(currentUser)){
+            throw new BizErrorException(ErrorCodeEnum.UAC10011039);
+        }
         Example example = new Example(SmtSupplier.class);
         example.createCriteria().andEqualTo("supplierCode",entity.getSupplierCode());
         SmtSupplier smtSupplier = smtSupplierMapper.selectOneByExample(example);
@@ -103,49 +101,59 @@ public class SmtSupplierServiceImpl  extends BaseService<SmtSupplier> implements
             throw new BizErrorException(ErrorCodeEnum.OPT20012001);
         }
         entity.setModifiedTime(new Date());
-//        entity.setModifiedUserId(currentUser.getUserId());
+        entity.setModifiedUserId(currentUser.getUserId());
         int i = smtSupplierMapper.updateByPrimaryKeySelective(entity);
 
         //获取当前供应商的地址集合对象
         List<SmtAddressDto> address = entity.getList();
 
-        //获取要删除的供应商地址
+        //获取当前要删除的供应商地址ID
         List<Long> delete = smtAddressMapper.findDelete(address, entity.getSupplierId());
-        smtAddressMapper.deleteByIds(org.apache.commons.lang3.StringUtils.join(delete,","));
+        if (delete.size()!=0){
+            smtAddressMapper.deleteByIds(org.apache.commons.lang3.StringUtils.join(delete,","));
+            //删除供应商与地址关联表中数据
+            Example example1 = new Example(SmtSupplierAddress.class);
+            example1.createCriteria().andIn("addressId",delete);
+            smtSupplierAddressMapper.deleteByExample(example1);
+        }
+
+        //获取当前供应商的所有地址集合
         List<SmtAddressDto> allAddress = smtAddressMapper.findBySupplierId(entity.getSupplierId());
+        List<SmtAddress> updateAddress = new ArrayList<>();
+        //批量修改地址，比较修改的地址对象
+        for (SmtAddressDto smtAddressDto : address) {
+            for (SmtAddressDto addressDto : allAddress) {
+                if (ClassCompareUtil.compareObject(smtAddressDto,addressDto)){
+                    smtAddressDto.setModifiedTime(new Date());
+                    smtAddressDto.setModifiedUserId(currentUser.getUserId());
+                    updateAddress.add(smtAddressDto);
+                }
+            }
+        }
+        if (updateAddress.size()!=0){
+            smtAddressMapper.batchUpdate(updateAddress);
+        }
 
-        smtAddressMapper.batchUpdate(address);
-
-//        Example example1 = new Example(SmtSupplier.class);
-//        example1.createCriteria().andEqualTo("supplierId",entity.getSupplierId());
-//        //查询当前供应商的地址关联关系对象集合
-//        List<SmtSupplierAddress> smtSupplierAddresses = smtSupplierAddressMapper.selectByExample(example1);
-//        //清空关联关系
-//        smtSupplierAddressMapper.deleteByExample(example1);
-
-//        List<SmtSupplierAddress> supplierAddresses = new ArrayList<>();
-//        List<SmtAddressDto> smtAddresses = new ArrayList<>();
-//        List<Long> deleteId = new ArrayList<>();
-//        //获取新增的地址对象
-//        for (SmtAddressDto smtAddress : address) {
-//            if (StringUtils.isEmpty(smtAddress.getAddressId())){
-//                smtAddresses.add(smtAddress);
-//            }else {
-//
-//            }
-//        }
-//
-//        for (SmtAddressDto smtAddress : address) {
-//            SmtSupplierAddress smtSupplierAddress = new SmtSupplierAddress();
-//            smtSupplierAddress.setAddressId(smtAddress.getAddressId());
-//            smtSupplierAddress.setSupplierId(entity.getSupplierId());
-//            smtSupplierAddress.setIfDefault(smtAddress.getIfDefault());
-//            supplierAddresses.add(smtSupplierAddress);
-//        }
-//
-//
-//
-//        smtSupplierAddressMapper.insertList(supplierAddresses);
+        List<SmtSupplierAddress> supplierAddresses = new ArrayList<>();
+        List<SmtAddressDto> smtAddresses = new ArrayList<>();
+        //获取新增的地址对象，并且保存
+        for (SmtAddressDto smtAddress : address) {
+            if (StringUtils.isEmpty(smtAddress.getAddressId())){
+                smtAddresses.add(smtAddress);
+            }
+        }
+        if (smtAddresses.size()!=0){
+            smtAddressMapper.insertList(smtAddresses);
+            //将新增地址与供应商关联，并且保存
+            for (SmtAddressDto smtAddress : smtAddresses) {
+                SmtSupplierAddress smtSupplierAddress = new SmtSupplierAddress();
+                smtSupplierAddress.setAddressId(smtAddress.getAddressId());
+                smtSupplierAddress.setSupplierId(entity.getSupplierId());
+                smtSupplierAddress.setIfDefault(smtAddress.getIfDefault());
+                supplierAddresses.add(smtSupplierAddress);
+            }
+            smtSupplierAddressMapper.insertList(supplierAddresses);
+        }
 
         return i;
     }
